% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_gwas}
\alias{build_gwas}
\title{Open a SNP GDS file and extract information.}
\usage{
build_gwas(gds_data = "name.gds", caco = "name.Rda", gwas_generator = TRUE)
}
\arguments{
\item{gds_data}{File name}

\item{caco}{An object of class caco. Output of \code{case_control} function.}

\item{gwas_generator}{logical; if \code{TRUE} an object of class result_snps
will be saved inside output list.}
}
\value{
a list of seven objects; including smp.id, snp.id, snp.pos, smp.indx,
smp.snp (a matrix with samples in rows and snp in columns), caco,
snps(column sum of smp.snp for each case control)
}
\description{
Open a SNP GDS file and extract information.
}
