% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_gwid}
\alias{build_gwid}
\title{Open a ibd file and extract information.}
\usage{
build_gwid(
  ibd_data = "name.ibd",
  gwas = "object of class gwas",
  gwid_generator = TRUE
)
}
\arguments{
\item{ibd_data}{a file name for output of \href{http://faculty.washington.edu/browning/refined-ibd.html}{Refined IBD}}

\item{gwas}{object of class gwas}

\item{gwid_generator}{logical; if \code{TRUE} an object of class result_snps
will be saved inside output list.}
}
\value{
the output will be a object(list) of class gwid contains
profile object, IBD object and result_snps object.
}
\description{
Open a ibd file and extract information.
}
