% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_contrl.R
\name{case_control}
\alias{case_control}
\title{Reload saved case-control list file}
\usage{
case_control(case_control_rda, ...)
}
\arguments{
\item{case_control_rda}{A character string giving the name of the case-control
file to load. The file is a list of character vectors including subject names
in each case-control groups or csv file including subject name for a disease.}

\item{...}{name of a column (disease name) of csv file.}
}
\value{
The output will be a list of character vectors include subject names
and groups. The class of returned object is caco.
}
\description{
Reload saved case-control list file
}
