% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract.gwas}
\alias{extract.gwas}
\title{Extract information from SNP GDS file.}
\usage{
\method{extract}{gwas}(obj, type = c("snps", "snp2", "nas"), snp_start, snp_end, ...)
}
\arguments{
\item{obj}{object of class gwas.}

\item{type}{indicate type of aggregation on sample-snp data and must be
one of snps, snp2, or nas}

\item{snp_start}{select starting position of snp, which we want to aggregate.}

\item{snp_end}{select ending position of snp, which we want to aggregate.}

\item{...}{other arguments}
}
\value{
the output will be a result_snps (data.table) object including 3 columns
including, snp_pos, case_control, and value
}
\description{
Extract information from SNP GDS file.
}
