% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract information from SNP GDS file.}
\usage{
extract(obj, ...)
}
\arguments{
\item{obj}{an object of class gwas}

\item{...}{other arguments}
}
\value{
extract object instants
}
\description{
Extract information from SNP GDS file.
}
