% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_window.gwid}
\alias{extract_window.gwid}
\title{Extract information from ibd data in a moving window}
\usage{
\method{extract_window}{gwid}(obj, w = 10, snp_start, snp_end, ...)
}
\arguments{
\item{obj}{object of class gwid(output of function build_gwid)}

\item{w}{window size}

\item{snp_start}{select starting position of snp, which we want to aggregate.}

\item{snp_end}{select ending position of snp, which we want to aggregate.}

\item{...}{other variables}
}
\value{
the output will be a result_snps (data.table) object including 3 columns
including, \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value}
}
\description{
Extract information from ibd data in a moving window
}
