% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_window}
\alias{extract_window}
\title{extract component of an object}
\usage{
extract_window(obj, ...)
}
\arguments{
\item{obj}{obj}

\item{...}{other variables}
}
\value{
the output will be a result_snps (data.table) object including 3 columns
including, \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value}
}
\description{
extract component of an object
}
