% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{fisher_test.gwid}
\alias{fisher_test.gwid}
\title{Fisher's Exact Test for gwid count data}
\usage{
\method{fisher_test}{gwid}(
  obj,
  caco,
  snp_start,
  snp_end,
  reference,
  alternative = c("two.sided", "greater", "less"),
  ...
)
}
\arguments{
\item{obj}{An object of class gwid. Output of \code{build_gwid} function}

\item{caco}{An object of class caco. Output of \code{case_control} function.}

\item{snp_start}{select starting position of snps.}

\item{snp_end}{select ending position of snp.}

\item{reference}{reference group of subjects in which we want to perform fisher test}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".
You can specify just the initial letter. Only used in the 2 by 2 case}

\item{...}{optional arguments to fisher.test}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
Fisher's Exact Test for gwid count data
}
