% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{fisher_test.result_snps}
\alias{fisher_test.result_snps}
\title{fisher exact test for result_snps count data}
\usage{
\method{fisher_test}{result_snps}(
  obj,
  caco,
  reference,
  alternative = c("two.sided", "greater", "less"),
  ...
)
}
\arguments{
\item{obj}{An object of class result_snps}

\item{caco}{An object of class caco. Output of \code{case_control} function.}

\item{reference}{reference group of subjects in which we want to perform fisher test.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".
You can specify just the initial letter. Only used in the 2 by 2 case}

\item{...}{optional arguments to fisher.test}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
fisher exact test for result_snps count data
}
