% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{gtest.haplotype_structure}
\alias{gtest.haplotype_structure}
\title{Perform G-test on haplotype structures extracted from \code{haplotype_structure} function}
\usage{
\method{gtest}{haplotype_structure}(haplotype_structure, reference, ...)
}
\arguments{
\item{haplotype_structure}{An object of class haplotype_structure. Output of
\code{haplotype_structure} function.}

\item{reference}{reference group of subjects in which we want to perform G-test}

\item{...}{other variables}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
Perform G-test on haplotype structures extracted from \code{haplotype_structure} function
}
