% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{gtest}
\alias{gtest}
\title{perform gtest}
\usage{
gtest(haplotype_structure, ...)
}
\arguments{
\item{haplotype_structure}{object of a class}

\item{...}{other variables}
}
\description{
perform gtest
}
