% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwid.R
\name{gwid}
\alias{gwid}
\title{Genome Wide Identity by Descent (GWID)}
\usage{
gwid(gds_data, caco, ibd_data, vcf_data, i, ...)
}
\arguments{
\item{gds_data}{CoreArray Genomic Data Structure (GDS) with hierarchical structure to store
multiple scalable structure array oriented data sets.}

\item{caco}{A List of case and control subjects.}

\item{ibd_data}{Data frame containing identity by descent segments in phased data. The data is output of refined_ibd
software package.}

\item{vcf_data}{Phased genotype data in vcf format.}

\item{i}{chromosome number can be from 1 to 22}

\item{...}{argument to pass to plots (if we decided to add graphs)}
}
\value{
data frames that are derived from inputs in order to overcome computational obstacle.
}
\description{
Genome Wide Identity by Descent (GWID)
}
