% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_structure.R
\name{haplotype_frequency.haplotype_structure}
\alias{haplotype_frequency.haplotype_structure}
\title{haplotype frequency in sliding windows}
\usage{
\method{haplotype_frequency}{haplotype_structure}(haplotype_structure, ...)
}
\arguments{
\item{haplotype_structure}{An object of class haplotype_structure. Output of
\code{haplotype_structure} function.}

\item{...}{other variables}
}
\value{
An object of class haplotype_frequency contains of two objects. first one
is object of haplotype_structure_frequency (data.table) and second one is object of class result_snps(data.table)
}
\description{
haplotype frequency in sliding windows
}
