% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_structure.R
\name{haplotype_frequency}
\alias{haplotype_frequency}
\title{haplotype frequency}
\usage{
haplotype_frequency(haplotype_structure, ...)
}
\arguments{
\item{haplotype_structure}{object of class haplotype structure}

\item{...}{other variables}
}
\description{
haplotype frequency
}
