% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_structure.R
\name{haplotype_structure.gwas}
\alias{haplotype_structure.gwas}
\title{extract haplotype structures of individuals in a window}
\usage{
\method{haplotype_structure}{gwas}(obj, phase, w = 10, snp_start, snp_end, ...)
}
\arguments{
\item{obj}{object of class gwas}

\item{phase}{An object of class phase. Output of \code{build_phase} function}

\item{w}{window size}

\item{snp_start}{select starting position of snps.}

\item{snp_end}{select ending position of snps.}

\item{...}{other variables}
}
\value{
The output will be an object of class haplotype_structure (data.table) that has information
about subjects haplotype structures in a a window.
}
\description{
extract haplotype structures of individuals in a window
}
