% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_structure.R
\name{haplotype_structure}
\alias{haplotype_structure}
\title{haplotype structures in a window}
\usage{
haplotype_structure(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other variables}
}
\description{
haplotype structures in a window
}
