% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{mcnemar_test}
\alias{mcnemar_test}
\title{mcnemar test}
\usage{
mcnemar_test(roh, ...)
}
\arguments{
\item{roh}{roh as class result_snp}

\item{...}{other variables}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
mcnemar test
}
