% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{mcnemar_test.result_snps}
\alias{mcnemar_test.result_snps}
\title{mcnemar test}
\usage{
\method{mcnemar_test}{result_snps}(
  roh = "object of class result_snps (output of function roh with fun=sum)",
  reference,
  w = 10,
  ...
)
}
\arguments{
\item{roh}{An object of class result_snps (output of function roh with fun=sum)}

\item{reference}{reference group of subjects in which we want to perform fisher test.}

\item{w}{window size}

\item{...}{other variables}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
mcnemar test
}
