% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{mcnemar_test_permut}
\alias{mcnemar_test_permut}
\title{mcnemar permutation}
\usage{
mcnemar_test_permut(mcnemar, ...)
}
\arguments{
\item{mcnemar}{macnemar test output}

\item{...}{other variables}
}
\description{
mcnemar permutation
}
