% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{mcnemar_test_permut.result_snps}
\alias{mcnemar_test_permut.result_snps}
\title{mcnemar permutation test}
\usage{
\method{mcnemar_test_permut}{result_snps}(
  mcnemar = "object of class result_snps (output of function mcnemar_test with fun=sum)",
  roh_mat = "output of roh function when roh_mat = TRUE",
  gwas = "object of class gwas",
  nperm = 1000,
  reference = "cases",
  w,
  ...
)
}
\arguments{
\item{mcnemar}{macnemar test output}

\item{roh_mat}{roh matrix}

\item{gwas}{gwas}

\item{nperm}{number of permutation}

\item{reference}{reference group}

\item{w}{window}

\item{...}{other variables}
}
\description{
mcnemar permutation test
}
