% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{permutation_test.gwas}
\alias{permutation_test.gwas}
\title{Permutation test for gwas object}
\usage{
\method{permutation_test}{gwas}(
  obj,
  snp_start,
  snp_end,
  nperm = 1000,
  reference = "cases",
  ...
)
}
\arguments{
\item{obj}{object of class gwas}

\item{snp_start}{elect starting position of snps.}

\item{snp_end}{select ending position of snp.}

\item{nperm}{Number of permutations.}

\item{reference}{reference group of subjects in which we want to perform fisher test}

\item{...}{other variables}
}
\description{
Permutation test for gwas object
}
