% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{permutation_test.gwid}
\alias{permutation_test.gwid}
\title{permutation test for gwid count data}
\usage{
\method{permutation_test}{gwid}(
  obj,
  gwas,
  snp_start,
  snp_end,
  nperm = 1000,
  reference = "cases",
  ibd_data,
  ...
)
}
\arguments{
\item{obj}{An object of class gwid. Output of \code{build_gwid} function}

\item{gwas}{object of class gwas}

\item{snp_start}{select starting position of snps.}

\item{snp_end}{select ending position of snp.}

\item{nperm}{Number of permutations.}

\item{reference}{reference group}

\item{ibd_data}{a file name for output of \href{http://faculty.washington.edu/browning/refined-ibd.html}{Refined IBD}}

\item{...}{other variables}
}
\description{
permutation test for gwid count data
}
