% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{permutation_test.haplotype_structure}
\alias{permutation_test.haplotype_structure}
\title{Permutation test for `haplotype_structure` object}
\usage{
\method{permutation_test}{haplotype_structure}(obj, nperm, reference, ...)
}
\arguments{
\item{obj}{object of class `haplotype_structure`}

\item{nperm}{Number of permutations.}

\item{reference}{reference group of subjects in which we want to perform `gtest`}

\item{...}{other variables}
}
\description{
Permutation test for `haplotype_structure` object
}
