% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{permutation_test}
\alias{permutation_test}
\title{permutation test}
\usage{
permutation_test(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other variables}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
permutation test
}
