% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gwas}
\alias{plot.gwas}
\title{Line plot of gwas objects}
\usage{
\method{plot}{gwas}(x, y = NA, title = "number of snps", ...)
}
\arguments{
\item{x}{object of class gwas.}

\item{y}{default value is NA, if specified it should be a vector of names of
subject groups i.e. y = c("case","control")}

\item{title}{title of the plot.}

\item{...}{optional argument of \code{plot}}
}
\value{
an interactive line plot of gwas objects for each case control subjects.
}
\description{
Line plot of gwas objects
}
