% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gwid}
\alias{plot.gwid}
\title{Line plot of gwid objects}
\usage{
\method{plot}{gwid}(
  x,
  y = NA,
  title = "number of IBD in each snp",
  plot_type = c("result_snps", "profile"),
  reference,
  ...
)
}
\arguments{
\item{x}{An object of class gwid. Output of \code{build_gwid} function.}

\item{y}{NA}

\item{title}{title of the plot.}

\item{plot_type}{either \dQuote{result_snps} or \dQuote{profile}.}

\item{reference}{reference group of subjects in which we want to have profile plot.}

\item{...}{if plot_type is \dQuote{result_snps} it is optional argument of \code{plot}.
if plot_type is \dQuote{profile} we can subset plot based on snp_start and snp_end locations.}
}
\value{
if plot_type is \dQuote{result_snps} an interactive line plot of result_snps for each case control subjects.
if plot_type is \dQuote{profile} an interactive profile plot of identity by descent subjects in subset of locations.
}
\description{
Line plot of gwid objects
}
