% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.haplotype_frequency}
\alias{plot.haplotype_frequency}
\title{Line plot of haplotype_frequency object}
\usage{
\method{plot}{haplotype_frequency}(
  x,
  y = NA,
  plot_type = c("haplotype_structure_frequency", "result_snps"),
  type = c("version1", "version2"),
  nwin,
  title,
  ...
)
}
\arguments{
\item{x}{an object of class haplotype_frequency}

\item{y}{NA}

\item{plot_type}{either \dQuote{result_snps} or \dQuote{"haplotype_structure_frequency"}}

\item{type}{either \dQuote{version1} or \dQuote{version2} when plot_type is \dQuote{"haplotype_structure_frequency"}}

\item{nwin}{window number}

\item{title}{title of the plot.}

\item{...}{optional argument of \code{plot}}
}
\description{
Line plot of haplotype_frequency object
}
