% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.result_snps}
\alias{plot.result_snps}
\title{Line plot of result_snps objects}
\usage{
\method{plot}{result_snps}(x, y = NA, title, snp_start, snp_end, ...)
}
\arguments{
\item{x}{An object of class result_snps.}

\item{y}{NA}

\item{title}{title of the plot.}

\item{snp_start}{select starting position of snps.}

\item{snp_end}{select ending position of snps.}

\item{...}{other variables}
}
\value{
an interactive line plot of result_snps for each case control subjects.
}
\description{
Line plot of result_snps objects
}
