% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.test_snps}
\alias{plot.test_snps}
\title{Line plot of test_snps objects}
\usage{
\method{plot}{test_snps}(x, y = NA, title, snp_start, snp_end, ...)
}
\arguments{
\item{x}{an object of class test_snps.}

\item{y}{NA}

\item{title}{title of the plot.}

\item{snp_start}{select starting position of snps.}

\item{snp_end}{select ending position of snps.}

\item{...}{other variables}
}
\value{
an interactive line plot of test_snps objects for each case control subjects.
}
\description{
Line plot of test_snps objects
}
