% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{roh.phase}
\alias{roh.phase}
\title{runs of homozygosity}
\usage{
\method{roh}{phase}(
  phase,
  gwas,
  w = 10,
  fun = c("sum", "mean"),
  snp_start,
  snp_end,
  roh_mat = FALSE,
  ...
)
}
\arguments{
\item{phase}{An object of class phase. Output of \code{build_phase} function}

\item{gwas}{object of class gwas}

\item{w}{window size}

\item{fun}{an aggregate function. either  \dQuote{sum} or  \dQuote{mean}}

\item{snp_start}{select starting position of snps.}

\item{snp_end}{select ending position of snps.}

\item{roh_mat}{return roh as matrix}

\item{...}{other variables}
}
\value{
the output will be a result_snps (data.table) object including 3 columns
including, \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value}
}
\description{
runs of homozygosity
}
