% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{roh}
\alias{roh}
\title{runs of homozygosity}
\usage{
roh(phase, ...)
}
\arguments{
\item{phase}{object of phase}

\item{...}{other variables}
}
\value{
runs of homozygosity data table or matrix
}
\description{
runs of homozygosity
}
