% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwid_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{run shiny app}
\usage{
run_shiny(data_folder_address, caco_file_name)
}
\arguments{
\item{data_folder_address}{Address for data folder.files with gds, ibd and
Rda format should named like this: chr1.gds, chr1.ibd, chr1.RDA}

\item{caco_file_name}{Case control file names. for example mycaco_RA.Rda}
}
\value{
shiny app
}
\description{
run shiny app
}
