% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{subset.gwid}
\alias{subset.gwid}
\title{subset gwid object based on snp position}
\usage{
\method{subset}{gwid}(obj, snp_start, snp_end, ...)
}
\arguments{
\item{obj}{object of class gwid(output of function build_gwid)}

\item{snp_start}{select starting position of snp, which we want to aggregate.}

\item{snp_end}{select ending position of snp, which we want to aggregate.}

\item{...}{other variables}
}
\value{
the output will be a object(list) of class gwid contains
profile object and result_snps object.
}
\description{
subset gwid object based on snp position
}
