% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{subset}
\alias{subset}
\title{subset an object}
\usage{
subset(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other variables}
}
\value{
the output will be a object(list) of class gwid contains
profile object and result_snps object.
}
\description{
subset an object
}
