% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_gwas}
\alias{build_gwas}
\title{Open a SNP GDS File and extract information.}
\usage{
build_gwas(gds_data = "name.gds", caco = "name.Rda", gwas_generator = TRUE)
}
\arguments{
\item{gds_data}{File name with \strong{gds} extension.}

\item{caco}{An object of class \emph{caco}. Output of \code{case_control} function.}

\item{gwas_generator}{logical; if \code{TRUE} an object of class \emph{result_snps}
will be saved inside output list.}
}
\value{
An object of \emph{gwas} class contains of seven objects; including smp.id, snp.id, snp.pos, smp.indx,
smp.snp (a matrix with samples in rows and snp in columns), caco,
snps(column sum of smp.snp for each case control)
}
\description{
This function allows the user to access genetic information stored in the \strong{gds} format,
which is the result produced by the \pkg{SNPRelate} package. In order to utilize this function,
the user must first employ \pkg{SNPRelate} to transform their file from \strong{vcf} format to the \strong{gds} format.
}
