% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_gwid}
\alias{build_gwid}
\title{Open an ibd file and extract information.}
\usage{
build_gwid(
  ibd_data = "name.ibd",
  gwas = "object of class gwas",
  gwid_generator = TRUE
)
}
\arguments{
\item{ibd_data}{a file name for output of \href{http://faculty.washington.edu/browning/refined-ibd.html}{Refined IBD}}

\item{gwas}{object of class \emph{gwas}}

\item{gwid_generator}{logical; if \code{TRUE} an object of class \emph{result_snps}
will be saved inside output list.}
}
\value{
the output will be a object of \emph{gwid} class contains
\emph{profile} object, \emph{IBD} object and \emph{result_snps} object.
}
\description{
The function \code{build_gwid()} processes an IBD file produced through \pkg{Refined IBD} software.
Subsequently, it constructs a \emph{gwid} class object comprising a sparse matrix of IBD segments. Moreover,
this \emph{gwid} object includes counts of IBD segments for individual pairs categorized within the case or control group.
}
