% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_phase}
\alias{build_phase}
\title{Read .vcf structured text format files and reduce the size of file.}
\usage{
build_phase(phased_vcf = "name.vcf", caco)
}
\arguments{
\item{phased_vcf}{A file name for a variant call format (vcf) file.}

\item{caco}{An object of class caco. Output of \code{case_control} function.}
}
\value{
the output will be a a list of class phase contains two sparse matrix
for each haplotype.
}
\description{
Read .vcf structured text format files and reduce the size of file.
}
