% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_structure.R
\name{haplotype_structure}
\alias{haplotype_structure}
\title{haplotype structures in a window}
\usage{
haplotype_structure(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other variables}
}
\value{
The output will be an object of class haplotype_structure (data.table) that has information
about subjects haplotype structures in a a window.
}
\description{
haplotype structures in a window
}
