% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{laucnh a shiny app}
\usage{
launch_app(data_folder_address, caco_file_name, ...)
}
\arguments{
\item{data_folder_address}{address of the folder that your data folders are.
for example if you have two sets of data such as data1 and data2 and they are
in mydata folder then your data_folder_address should be "./mydata"}

\item{caco_file_name}{case control file name.}

\item{...}{other variables}
}
\value{
open a shiny app
}
\description{
laucnh a shiny app
}
