% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roh.R
\name{mcnemar_test_permut}
\alias{mcnemar_test_permut}
\title{mcnemar permutation}
\usage{
mcnemar_test_permut(mcnemar, ...)
}
\arguments{
\item{mcnemar}{macnemar test output}

\item{...}{other variables}
}
\value{
the output will be a test_snps (data.table) object including 3 columns:
 \dQuote{snp_pos}, \dQuote{case_control}, and \dQuote{value} which is
a p-values.
}
\description{
mcnemar permutation
}
