% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{permutation_test}
\alias{permutation_test}
\title{permutation test}
\usage{
permutation_test(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other variables}
}
\description{
permutation test
}
