% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.haplotype_frequency}
\alias{plot.haplotype_frequency}
\title{Line plot of haplotype_frequency object}
\usage{
\method{plot}{haplotype_frequency}(
  x,
  y = NA,
  plot_type = c("haplotype_structure_frequency", "result_snps"),
  type = c("version1", "version2"),
  ly,
  nwin,
  title,
  ...
)
}
\arguments{
\item{x}{an object of class haplotype_frequency}

\item{y}{default value is `NA`, if specified it should be a vector of names of
subject groups i.e. `y = c("case","control")`}

\item{plot_type}{either \dQuote{result_snps} or \dQuote{"haplotype_structure_frequency"}}

\item{type}{either \dQuote{version1} or \dQuote{version2} when plot_type is \dQuote{"haplotype_structure_frequency"}}

\item{ly}{if TRUE, we have a plotly object and if it is false plot is going to be
a ggplot object.}

\item{nwin}{window number}

\item{title}{title of the plot.}

\item{...}{optional argument of \code{plot}}
}
\description{
Line plot of haplotype_frequency object
}
