% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.haplotype_structure_frequency}
\alias{plot.haplotype_structure_frequency}
\title{Two type of line plots for haplotype_structure_frequency objects.}
\usage{
\method{plot}{haplotype_structure_frequency}(x, y = NA, type = c("version1", "version2"), nwin, ...)
}
\arguments{
\item{x}{an object of class haplotype_structure_frequency}

\item{y}{NA}

\item{type}{either \dQuote{version1} or \dQuote{version2}}

\item{nwin}{window number}

\item{...}{other variables}
}
\description{
Two type of line plots for haplotype_structure_frequency objects.
}
